+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Criado ícones `BlockIcon`, `MessageIcon` e `TrashIcon`.
 - Criado componente `DeleteAccountCard` [#236] (https://github.com/3C-gg/reload-frontend/issues/236).
 - Adicionado método `delete` na interface de AccountsAPI.
 - Adicionado método `updateIsActive` na interface de AccountsAPI.
@@ -54,6 +55,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Alterado view de `Account` para refletir o design [#134](https://github.com/3C-gg/reload-frontend/issues/134).
+- Movido `HeaderProfile` e função `renderButtonsNavigation` para `LayoutProfile`.
+- Removido `HeaderProfile` e função `renderButtonsNavigation` da view `Profile`.
+- Alterado path da `AccountView` para `/conta/:userId`.
 - Alterado componente `ChangeEmailCard` para usar o método `updateEmail`.
 - Alterado background do `Modal` para `gray.800` no tema do chakra.
 - Alterado gray 800 para `#1B1B1B`.